feat: (#25) 페이지에서 노션 리스트의 노션 컴포넌트와 더하기 버튼을 가지도록 수정
